Use "long" for the content-length.
We should use "long" for the content-length, as the integer type in
Elasticsearch is limited to signed 32-bit integers, and fails for
objects greater than ~2GB in size.

+++ swift_metadata_sync/metadata_sync.py
@@ -13,7 +13,7 @@ from container_crawler.base_sync import BaseSync
 class MetadataSync(BaseSync):
     DOC_TYPE = 'object'
     DOC_MAPPING = {
-        "content-length": {"type": "integer"},
+        "content-length": {"type": "long"},
         "content-type": {"type": "string"},
         "etag": {"type": "string", "index": "not_analyzed"},
         "last-modified": {"type": "date"},
